pocs/tennis_coaching_team/requirements.txt
anthropic>=0.40.0
streamlit>=1.30.0
python-dotenv
